// src/app/service/model/CategoryDTO.ts
/**
 * SpongeblogSP API
 * Spongeblog blogging API
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface CategoryDTO {
    /**
     * Unique identifier.
     */
    categoryId?: number;

    links?: Array<models.Link>;

    /**
     * Category display name.
     */
    name?: string;

    /**
     * Child categories.
    */
    children?: Array<models.CategoryDTO>;

    /**
     * Id of parent Category.
     */
    parentId?: number;

    /**
     * Determines position of Category.
     */
    rank?: number;

    /**
     * ID of the assigned page if it's a page-category (optional field).
    */
    pageId?: number;
}