// src/app/service/model/TagDTO.ts
/**
 * SpongeblogSP API
 * Spongeblog blogging API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface TagDTO {
    links?: Array<models.Link>;

    /**
     * Tag name.
     */
    name?: string;

    /**
     * Unique identifier.
     */
    tagId?: number;

}